@@ -174,6 +174,7 @@
     "glob": "^7.1.1",
     "graceful-fs": "^4.1.3",
     "inquirer": "^3.0.6",
+    "invariant": "^2.2.4",
     "lodash": "^4.17.5",
     "metro": "^0.45.0",
     "metro-babel-register": "^0.45.0",
@@ -197,6 +198,7 @@
     "react-timer-mixin": "^0.13.2",
     "regenerator-runtime": "^0.11.0",
     "rimraf": "^2.5.4",
+    "scheduler": "^0.11.2",
     "semver": "^5.0.3",
     "serve-static": "^1.13.2",
     "shell-quote": "1.6.1",
@@ -222,7 +224,7 @@
     "jest": "22.0.0",
     "polished": "^1.9.3",
     "prettier": "^1.14.0",
-    "react": "16.2.0",
+    "react": "16.6.3",
     "react-test-renderer": "16.2.0",
     "shelljs": "^0.7.8",
     "sinon": "^2.2.0"
