@@ -1,20 +1,21 @@
 /**
- * Copyright (c) 2014-present, Facebook, Inc.
+ * Copyright (c) Facebook, Inc. and its affiliates.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
  * @flow
- * @providesModule ReactTypes
  */
 
 export type ReactNode =
   | React$Element<any>
-  | ReactCall
-  | ReactReturn
   | ReactPortal
   | ReactText
-  | ReactFragment;
+  | ReactFragment
+  | ReactProvider<any>
+  | ReactConsumer<any>;
+
+export type ReactEmpty = null | void | boolean;
 
 export type ReactFragment = ReactEmpty | Iterable<React$Node>;
 
@@ -22,20 +23,47 @@ export type ReactNodeList = ReactEmpty | React$Node;
 
 export type ReactText = string | number;
 
-export type ReactEmpty = null | void | boolean;
-
-export type ReactCall = {
+export type ReactProvider<T> = {
   $$typeof: Symbol | number,
+  type: ReactProviderType<T>,
   key: null | string,
-  children: any,
-  // This should be a more specific CallHandler
-  handler: (props: any, returns: Array<mixed>) => ReactNodeList,
-  props: any,
+  ref: null,
+  props: {
+    value: T,
+    children?: ReactNodeList,
+  },
 };
 
-export type ReactReturn = {
+export type ReactProviderType<T> = {
   $$typeof: Symbol | number,
-  value: mixed,
+  _context: ReactContext<T>,
+};
+
+export type ReactConsumer<T> = {
+  $$typeof: Symbol | number,
+  type: ReactContext<T>,
+  key: null | string,
+  ref: null,
+  props: {
+    children: (value: T) => ReactNodeList,
+    unstable_observedBits?: number,
+  },
+};
+
+export type ReactContext<T> = {
+  $$typeof: Symbol | number,
+  Consumer: ReactContext<T>,
+  Provider: ReactProviderType<T>,
+
+  _calculateChangedBits: ((a: T, b: T) => number) | null,
+
+  _currentValue: T,
+  _currentValue2: T,
+  _threadCount: number,
+
+  // DEV only
+  _currentRenderer?: Object | null,
+  _currentRenderer2?: Object | null,
 };
 
 export type ReactPortal = {
@@ -46,3 +74,7 @@ export type ReactPortal = {
   // TODO: figure out the API for cross-renderer implementation.
   implementation: any,
 };
+
+export type RefObject = {|
+  current: any,
+|};
